move calender, languageGroup and title into own component folder called common. Renamed to Title and LanguageGroup for better naming
