+++ todo-app-standalone/src/components/TaskAllTasksList.vue
@@ -47,7 +47,7 @@ import { userStore } from '../stores/userStore'
 
 import TextField from "@/components/form/TextField.vue"
 import CheckBox from "@/components/form/CheckBox.vue"
-import Calender from "@/components/Calender.vue"
+import Calender from "@/components/common/Calender.vue"
 import IconButton from "@/components/buttons/IconButton.vue";
 
 export default {
